backend/requirements.txt
# Core
flask>=3.0.0
flask-cors>=4.0.0
celery>=5.3.0
redis>=5.0.0
gunicorn>=21.0.0

# AI/ML
torch>=2.1.0
git+https://github.com/huggingface/diffusers.git
transformers
accelerate
safetensors
sentencepiece
peft
torchao==0.11.0
torchvision
rembg>=2.0.0

# TripoSR (3D Reconstruction)
torchmcubes
einops
omegaconf
huggingface_hub

# Mesh
trimesh>=4.0.0
open3d>=0.18.0

# Image
pillow>=10.0.0

# Utils
pydantic>=2.5.0
python-dotenv>=1.0.0
